resolve #5 verify client returns scopes
The authorization server may modify the lists of scopes allowed by the calling
client so we need to pass back this modified list so it may be used in
subsequent method calls.

note that this will *break* the Mojolicious plugin as it does not expect the
scopes ref to be returned.