Allow access_token_ttl to be passed as callback
In some scenarios we might have some specific business logic where TTL
depends on the client or other similar conditions.

To support client specific TTL allow `access_token_ttl` to be passed as
callback so that we can call our own conditions on what the lifespan of
the token should be.